pyproject.toml
[tool.poetry]
name = "individual work is for web_programming"
version = "0.1.0"
description = "This program is for user of tasks "
authors = ["Your Name thmzliorxan40@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
fastapi="0.111.0"
uvicorn="0.29.0"
pydantic="2.7.1"
python-multipart="0.0.9"
python-jose="3.3.0"
passlib="1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
